requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
asyncpg==0.29.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg[binary]==3.2.3
psycopg2-binary==2.9.9  # For benchmarks and partition management

# Redis
redis==5.0.1
hiredis==2.2.3

# LLM
openai==1.3.7

# Data Processing
pandas==2.1.3
numpy==1.26.2
openpyxl==3.1.2
scipy==1.11.4

# Visualization & Dashboard
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.18.0
streamlit==1.29.0

# Machine Learning (for segmentation)
scikit-learn==1.3.2

# Orchestration
prefect>=2.14.0,<3.0.0

# Data Quality
great-expectations==0.18.8

# dbt
dbt-core==1.7.0
dbt-postgres==1.7.0

# Utilities
python-dotenv==1.0.0
python-multipart==0.0.6
httpx==0.25.2

# Monitoring
structlog==23.2.0
prometheus-client==0.19.0

# Linting
sqlfluff==3.0.0
sqlfluff-templater-dbt==3.0.0
ruff==0.1.9
black==23.12.1
